refactor (move calculateWinPercent)
